[APM] Diagnostics: Communicate that CCS is not supported
