+++ x-pack/plugins/apm/public/components/app/diagnostics/apm_documents_tab.tsx
@@ -22,6 +22,14 @@ import { useDiagnosticsContext } from './context/use_diagnostics';
 import { ApmPluginStartDeps } from '../../../plugin';
 import { SearchBar } from '../../shared/search_bar/search_bar';
 
+function formatDocCount(count?: number) {
+  if (count === undefined) {
+    return '-';
+  }
+
+  return asInteger(count);
+}
+
 export function DiagnosticsApmDocuments() {
   const { diagnosticsBundle, isImported } = useDiagnosticsContext();
   const { discover } = useKibana<ApmPluginStartDeps>().services;
@@ -48,24 +56,24 @@ export function DiagnosticsApmDocuments() {
       name: '1m',
       field: 'intervals.1m',
       render: (_, { intervals }) => {
-        const interval = intervals?.['1m'];
-        return interval ? asInteger(interval) : '-';
+        const docCount = intervals?.['1m'];
+        return formatDocCount(docCount);
       },
     },
     {
       name: '10m',
       field: 'intervals.10m',
       render: (_, { intervals }) => {
-        const interval = intervals?.['10m'];
-        return interval ? asInteger(interval) : '-';
+        const docCount = intervals?.['10m'];
+        return formatDocCount(docCount);
       },
     },
     {
       name: '60m',
       field: 'intervals.60m',
       render: (_, { intervals }) => {
-        const interval = intervals?.['60m'];
-        return interval ? asInteger(interval) : '-';
+        const docCount = intervals?.['60m'];
+        return formatDocCount(docCount);
       },
     },
     {
