+++ x-pack/plugins/apm/public/components/app/diagnostics/summary_tab/index.tsx
@@ -6,19 +6,41 @@
  */
 
 import React from 'react';
-import { EuiFlexGroup } from '@elastic/eui';
+import { EuiFlexGroup, EuiCallOut, EuiSpacer } from '@elastic/eui';
 import { ApmIntegrationPackageStatus } from './apm_integration_package_status';
 import { IndexTemplatesStatus } from './index_templates_status';
 import { FieldMappingStatus } from './indicies_status';
 import { DataStreamsStatus } from './data_streams_status';
+import { useDiagnosticsContext } from '../context/use_diagnostics';
 
 export function DiagnosticsSummary() {
+  const { diagnosticsBundle, status, isImported } = useDiagnosticsContext();
+
+  const isCrossCluster = Object.values(
+    diagnosticsBundle?.apmIndices ?? {}
+  ).some((indicies) => indicies.includes(':'));
+
   return (
-    <EuiFlexGroup direction="column">
-      <ApmIntegrationPackageStatus />
-      <IndexTemplatesStatus />
-      <DataStreamsStatus />
-      <FieldMappingStatus />
-    </EuiFlexGroup>
+    <>
+      {isCrossCluster && (
+        <>
+          <EuiCallOut
+            title="Cross cluster search not supported"
+            color="warning"
+          >
+            The APM index settings is targetting remote clusters. Please note:
+            this is not currently supported by the Diagnostics Tool and
+            functionality will therefore be limited.
+          </EuiCallOut>
+          <EuiSpacer />
+        </>
+      )}
+      <EuiFlexGroup direction="column">
+        <ApmIntegrationPackageStatus />
+        <IndexTemplatesStatus />
+        <DataStreamsStatus />
+        <FieldMappingStatus />
+      </EuiFlexGroup>
+    </>
   );
 }
